refactor: OperatorType -> Operator로 변경, 책임 부여와 예외발생
OperatorType이었던 enum 클래스를 Operator로 변경후 기존의 Operator클래스를 제거했습니다. 그리고 연산자
스택을 관리하도록 책임을 부여했습니다. 그리고 예외를 발생하도록 코드를 변경했습니다.
